requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0.0
Flask-Assets==0.11
Flask-Admin==1.3.0
Flask-Login==0.3.2
Flask-WTF==0.12
raven[flask]==5.8.1
